migrate Semigroup and Monoid instances to Alt and Plus
Behavior can have Monoid instance since value is available at all time when
sampled by some Event.
But Event shouldn't have Monoid instance otherwise it becomes stateful, in the
sense that one Event has to hold a value somewhere to wait for a time in the
future to coordinate with the other Event.